requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.2.0
camel-kenlm==2021.12.27
camel-tools @ file:///home/nabeeli/camel_tools_github/camel_tools
certifi==2022.9.24
charset-normalizer==2.1.1
dill==0.3.5.1
Django==4.1.1
docopt==0.6.2
editdistance==0.6.0
emoji==2.1.0
filelock==3.8.0
future==0.18.2
huggingface-hub==0.10.0
idna==3.4
joblib==1.2.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sqlparse==0.4.3
tabulate==0.8.10
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.3.0
urllib3==1.26.12
